Handling for Paralax selection from dropdown
